refactor lease contract
